modify unittest for previous commit and some fix
fix at NetworkGateway:
- add return code if resource_id does not exist in handle_delete()

+++ heat/engine/resources/neutron/network_gateway.py
@@ -65,6 +65,8 @@ class NetworkGateway(neutron.NeutronResource):
             self.resource_id)['network_gateway']
 
     def handle_delete(self):
+        if not self.resource_id:
+            return
         client = self.neutron()
         try:
             client.delete_network_gateway(self.resource_id)
